requirements.txt
torch==1.3.0
numpy==1.17.2
torchvision==0.4.1a0+d94043a
Pillow==6.2.1
